add custom virus definition file to catch the eicar test signature as a substring
clamav's main definitions can't be relied upon to keep reacting to eicar in
the same way, and recently seem to have stopped detecting it when used as
a file substring. so we add a supplementary definition file to ensure our
smoulder tests will behave as we expect. definition file is stored in two
parts XOR'ed together as it contains a representation of the signature
itself and we don't want to trigger naive scanners everywhere we go with
this repo.

for bonus points, we could assemble the file at runtime instead of build
time so we don't store the final string in our docker images either